requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.31.0
yfinance==0.2.36 